backend/requirements.txt
fastapi
uvicorn
motor
pymongo
httpx
google-genai
python-dotenv
pydantic
